Move accounts-ldap out of package and use ldapjs from NPM
We can remove and simplify the LDAP integration code.  I'm using the precise
version specified in the previous package, rather than the current ldapjs
release, to minimize potential for breakage.  We can upgrade dependencies
another time.

This adds another ignorable warning to the console in dev mode about the "jsv"
package, which is an optional, dynamically-require()'d package from a
transitive dependency of ldapjs.
