+++ shell/.meteor/versions
@@ -1,6 +1,5 @@
 accounts-base@1.2.8
 accounts-identity@0.1.0
-accounts-ldap@1.0.1
 accounts-oauth@1.1.13
 accounts-saml@0.2.0
 allow-deny@1.0.5
@@ -100,7 +99,6 @@ standard-minifier-js@1.0.7
 templating@1.1.11
 templating-tools@1.0.4
 tracker@1.0.14
-typ:ldapjs@0.7.3
 ui@1.0.11
 underscore@1.0.9
 url@1.0.10
