+++ shell/server/accounts/ldap/LICENSE
